fix: Cognitive layer walkthrough #9 — 824 bugs across 540 files
Systemic fixes: nimcp_mutex_free→destroy (87), mutex cast removal (19),
dead (void)agent cleanup (~200), error code inversions (~50), NULL
heartbeat fixes (130+), thalamic bridge locking (15), callbacks under
mutex deadlock prevention, shallow copy aliasing fixes.

Critical: executive mutual recursion (stack overflow), VAE self-deadlock,
mental health guardian spurious throw, creative subsystem heap corruption/
OOB/double-free.

+++ src/cognitive/creative/external/nimcp_diffusion_bridge.c
@@ -403,7 +403,7 @@ int diffusion_text_to_image(diffusion_bridge_t* bridge,
     output->width = width;
     output->height = height;
     output->channels = 3;
-    output->pixels = nimcp_calloc(width * height * 3, sizeof(uint8_t));
+    output->pixels = nimcp_calloc((size_t)width * (size_t)height * 3, sizeof(uint8_t));
     if (!output->pixels) {
         set_diffusion_error("Failed to allocate output image");
         NIMCP_THROW_TO_IMMUNE(NIMCP_ERROR_NO_MEMORY, "diffusion_text_to_image: output->pixels is NULL");
